download fedora x86_64 image if heat is enabled
clean up the logic around producing a valid cfn image for heat
testing. If heat is enabled, add the f20 x86_64 imageurl to the
list of images that we support, and set that as the tempest
orchestration image.

The crux of this change is left commented out for now because we
need a two phase update of nodepool content before we can move
away from the old image (otherwise fedora mirrors will kill us
with bad downloads).

Change-Id: I1da9f259a3267b1bd27e9899755204c8fff41153