Inject the domain name for devstack in tempest
A new setting for the keystone domain name is introduced in
tempest by change https://review.openstack.org/#/c/81872/.

The default value in tempest is set to None, so that it is
not devstack specific. It's up to devstack to configure
the name of the keystone domain defined in devstack.

Partially implements bp multi-keystone-api-version-tests

Change-Id: I3d1406407759e1fdce0704fa2d6aa025f23a5f08