Kill spinner process when stack.sh exits
The last spinner process active in non-verbose mode does not get killed
when stack.sh exits -- the spinner keeps spinning indefinitely.

Killing the spinner in err_exit cleans up no matter how the program
got terminated.

Because the code to kill the spinner is now called regardless of
whether spinners are in use, it has to check LAST_SPINNER_PID or the
kill command without an argument will trigger the ERR trap (or EXIT
with an error status, depending on where program execution stops).

This patch resurrects and fixes an abandoned changeset, hence:

Co-Authored-By: Adalberto Medeiros <adalbas@linux.vnet.ibm.com>

Fixes bug 1302112

Change-Id: I2d5b27971889b672361e9173bf6faf38fb1a1ec6