Remove copy of cinder.conf.sample file
Rather than continue to be bit by conf updates
from external sources, Cinder is going to join
the party and cease generation of the sample.conf
as part of the unit tests.

First step however is to not attempt to copy it
from cinder/etc/cinder in the devstack setup.

This just removes that cp operation, nothing
in that sample conf was modfied from defaults
so it wasn't really doing anything anyway.

Change-Id: I727b1747ff36d775ef881e93e3b9a8528a17a4df