Cut the heat flavor
1024Mb memory is too much for single instance.
approximately 4 GiB memory used even without a booted VM.

The heat auto-scaling scenario needs to boot 3 VM
(currently disabled), if any other test running in parallel
and boots up a vm it would easily can cause OOM an
issue.

The heat-slow job can work with 128 MiB instance memory,
but this change reduces it just to 512 MiB (safer for the guest),
and enough to avoid OOM issues on the host,
and let us to heve more vm at the same time.

The m1.tiny flavor is not good because now it has a hard coded
1 GiB disk limit and the f20 image requires 2 GiB.
Setting the disk size to '0' -> automatically use the virtual size
of the image.

Using multiple cores does not really helps in boot time :(,
but it might cause higher host load. (More locking)
Reducing the core number to 1.

Change-Id: Ice7c1253d15e91fd1c02bc8c6809335aaebfff4a